+++ src/main/java/com/epam/lab/news/config/MongoDbConfig.java
@@ -2,8 +2,15 @@ package com.epam.lab.news.config;
 
 import com.mongodb.Mongo;
 import com.mongodb.MongoClient;
+import com.mongodb.MongoClientOptions;
+import com.mongodb.ServerAddress;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.mongodb.MongoDbFactory;
 import org.springframework.data.mongodb.config.AbstractMongoConfiguration;
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.SimpleMongoDbFactory;
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 
 /**
  * Defines MongoDB configuration
@@ -12,6 +19,7 @@ import org.springframework.data.mongodb.config.AbstractMongoConfiguration;
  * @since 0.1.0-alpha
  */
 @Configuration
+@EnableMongoRepositories(basePackages = "com.epam.lab.news.data.repo")
 public class MongoDbConfig extends AbstractMongoConfiguration {
     /** Keeps database name */
     private static final String DATABASE = "news";
@@ -33,15 +41,33 @@ public class MongoDbConfig extends AbstractMongoConfiguration {
         return DATABASE;
     }
 
+    @Bean
+    public MongoTemplate mongoTemplate() throws Exception {
+        return new MongoTemplate(mongoDbFactory());
+    }
+
+    @Bean
+    public MongoDbFactory mongoDbFactory() throws Exception {
+        return new SimpleMongoDbFactory(mongo(), DATABASE);
+    }
+
     /**
-     * Returns Mongo object which contains internal pool with 100 connections.
+     * Returns Mongo object initialized by custom pool options.
      *
      * @return Mongo object
      * @throws Exception if database cannot be resolved
      */
     @Override
     public Mongo mongo() throws Exception {
-        return new MongoClient(HOST, PORT);
+        MongoClientOptions options = MongoClientOptions.builder()
+                .connectionsPerHost(8)
+                .threadsAllowedToBlockForConnectionMultiplier(4)
+                .connectTimeout(100)
+                .maxWaitTime(500)
+                .socketKeepAlive(true)
+                .socketTimeout(500)
+                .build();
+        return new MongoClient(new ServerAddress(HOST, PORT), options);
     }
 
 }