+++ src/main/java/com/epam/lab/news/data/repo/NewsRepository.java
@@ -1,7 +1,6 @@
 package com.epam.lab.news.data.repo;
 
 import com.epam.lab.news.bean.Article;
-import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -13,6 +12,5 @@ import org.springframework.stereotype.Repository;
  * @since 1.0.1-alpha
  */
 @Repository
-@EnableMongoRepositories
 public interface NewsRepository extends CrudRepository<Article, Long> {
 }